pkvm: x86: Introduce def.h to control defines for pkvm hypervisor
Introduce a def.h to control defines for pkvm hypervisor. It is included
on top of each pkvm hypervisor source code so that the defines in def.h
can take effact at the beginning. This could be used by pkvm to undefine
some unnecessary linux kernel configs to avoid introducing the unneeded
symbols by using linux kernel macros, or re-define a linux kernel macro
with pkvm's own implementation.

The kernel config CONFIG_DEBUG_PREEMPT is undefined. This is used for
the smp_processor_id() macro defined in linux kernel but is unnecessary
when it is used by the pkvm hypervisor to get the CPU id.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>