pkvm: vmx: Handle EPT_MISCONFIG vmexit
For the pVM, it will be enlighted to use hypercall to access MMIO thus
there is no EPT_MISCONFIG. Inject #GP to the guest if this happens.

For the npVM, the EPT_MISCONFIG is allowed. Share the GPA information
with the host via the exit_gpa field in the vcpu_vmx structure.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>