pkvm: vmx: Handle IO vmexit
For the pVM, only non-string IO vmexit is allowed for the host KVM to
emulate. The string IO vmexit requires instruction decoding and
emulating, so it is not supported. The pkvm hypervisor will inject #GP
to the guest if the string IO vmexit happens. The necessary pVM vcpu
registers are shared to the host KVM to emulate the IO instruction. The
pkvm hypervisor will update the guest registers accordingly before
entering the guest.

For the npVM, both non-string/string IO vmexits are supported.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>