pkvm: vmx: Pass vm/vcpu handle for PKVM_HC_SET_MMIO_VE
The original PKVM_HC_SET_MMIO_VE implementation relies on the pkvm
hypervisor to record a current_shadow_vcpu pointer when emulates the
vmptrld instruction. This blocks to remove the VMX emulation code. To
resolve this, passing the vm/vcpu handles for this hypercall so that the
pkvm hypervisor can get the corresponding kvm_vcpu via the handles
rather than using the current_shadow_vcpu pointer.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>