pkvm: x86: Handle the KVM_CALL from the linux host
Introduce a pkvm.c in kvm/pkvm/ which acts as a x86 common code for both
VMX and SVM. The KVM_CALL will be used to implement the PV ABI which
hooks in the kvm_x86_ops. As kvm_x86_ops has common meaning for both VMX
and SVM, implementing the KVM_CALL handler in the pkvm x86 common file.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>