pkvm: x86: Add struct pkvm_vm and pkvm_vm_ref
The struct pkvm_vm represents a vm structure in the pkvm hypervisor. It
includes a pointer pointing to the struct kvm in the linux host, its
size and other elements for the vm management.

The struct kvm can be appended in the end of pkvm_vm but not explicitly
included as below:
  -------------------
  | struct pkvm_vm  |
  -------------------
  | struct kvm      |
  -------------------

The reason of *NOT* explicitly putting struct kvm inside the pkvm_vm is
that the struct kvm may be wrapped by the architecture specific
structure, e.g. struct kvm_vmx. Appending struct kvm in the end is
friendly to this case as the architecture specific structure can be
appended in the end if the struct kvm is wrapped at the offset 0. The
helper function to_kvm can be used to convert struct pkvm_vm * to struct
kvm *. So the actual pkvm_vm size may depends on the architecture.
Introduce the global pkvm_vm_sz for the architecture code to override.

The struct pkvm_vm_ref is used to hold the reference count for each
created pkvm_vm, which is facilitate for vm life cycle management.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>