pkvm: x86: Add vm_init/destroy PV interfaces
Add PV interfaces for linux host to do the vm init and destroy in the
pkvm hypervisor. The vm_init responds for donate the memory pages and
initialize the pkvm_vm structure. The architecture related vm
initialization is done via the architecture kvm_x86_ops. The vm_destroy
is basically doing the opposite.

The inputs for vm_init are the pointer pointing to the struct kvm in
linux host, and a physical address allocated by linux host for the pkvm
hypervisor to creat pkvm_vm structure. The output is the vm handle if
success, otherwise the error code.

The input for vm_destroy is the vm handle and the output is the value
indicating if success or not.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>