pkvm: x86: Add vcpu_create PV interface
Add PV interface for the linux host to do vcpu_create in the pkvm
hypervisor. The vcpu_create is implemented as a pkvm common function
which donates the pages from the linux host and uses as the pkvm_vcpu
structure. The architecture specific vcpu initialization is done via the
kvm_x86_ops.vcpu_create.

The reason for not having a PV interface for vcpu_free is that the pkvm
vcpu related pages are undonated by pkvm hypervisor together when
destroy the VM, which is simpler and aligned with what pkvm-arm did.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>