kvm: pkvm: Disallow the host to change MSR exit bitmap
Whether accessing a MSR in the guest causing vmexit or not is controlled
by the pkvm hypervisor via the MSR exit bitmap. The host is not allowed
to modify this so there is no PV interface exposed. The MSR filter
configured by the host will be applied when the host emulates a MSR.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>