pkvm: vmx: Implement prepare_switch_to_guest/host
Add a new callback in kvm_x86_ops as prepare_switch_to_host, to pair
with the existing callback prepare_switch_to_guest. The implementation
for these two functions are referenced from the linux VMX code. The uret
MSRs switching is not included as the corresponding emulation are still
done by the host KVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>