REVERTME: pkvm: vmx: Sync irq/nmi/exception related fields
Sync the irq/nmi/exception related fields between the shared kvm_vcpu
and private kvm_vcpu so that both sides can trigger event injection.

In the future, the irq/nmi injection will be triggered by the host KVM
and the exception injection will be triggered by the pkvm hypervisor.
With this, the sync will not be required.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>