pkvm: x86: Add vcpu_load PV interface
Add vcpu_load PV interface for the host KVM to load a vcpu in the pkvm
hypervisor. The loaded vcpu will be set as inuse so that the host KVM
cannot ask pkvm hypervisor to arbitrary destroy this vm while pkvm
hypervisor is still using its vcpu. The inuse state will be reverted by
the vcpu_put PV interface, which is implemented by the following patch.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>