pkvm: x86: Implement kvm_emulate_rdmsr/wrmsr
Implement the kvm_emulate_rdmsr/wrmsr so that the MSR vmexit can be
handled by the pkvm hypervisor in the first place. As not all MSRs
emulation are supported in the pkvm hypervisor, the unsupported MSRs
will result in switching back to the host KVM to emulate.

For the supported MSR emulation, the pkvm hypervisor can skip the MSR
instruction if the emulation is successful. Otherwise, a #GP exception
should be injected to the guest. As injecting an exception is not
supported yet, currently the pkvm hypervisor will switch back to the
host KVM to do so.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>