pkvm: vmx: Implement set/get_msr
Implement the set/get_msr callbacks in the vt_x86_ops to emulate the MSR
reading and writing. With this, most of the intel MSRs emulation are
supported in the pkvm hypervisor, including the uret MSRs, except for
the MSRs related to the nested and intel processor trace. For the MSRs
which can be emulated by the kvm_set/get_msr_common are also not
implemented yet.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>