pkvm: x86: Add vcpu_after_set_cpuid PV interface
Add the vcpu_after_set_cpuid PV interface to copy the cpuid entries from
the host KVM to the pkvm hypervisor.

The host KVM allocates pages for the cpuid entries in the pkvm
hypervisor. These pages will be donated to the pkvm hypervisor by this
PV interface, and undonated when VM is destroyed. The pkvm hypervisor
will use kvm_set_cpuid() to copy the cpuid entries from the host KVM to
these pages. If the pkvm hypervisor already have cpuid entries, the old
pages will be undonated and its physical address will be returned back
to the host KVM to free.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>