kvm: pkvm: Use PV interface to do vcpu reset
Add vcpu_reset PV interface for the host KVM to reset the vcpu in the
pkvm hypervisor. As the host KVM still emulates the virtual lapic, the
cr8 and lapic are still handled by the host KVM.

Once the guest state is protected, resetting vcpu is not allowed via
this PV interface.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>