kvm: pkvm: Use PV interface to access idt/gdt
Implement set/get_idt/gdt in vt_x86_ops for the pkvm hypervisor by
porting from Linux VMX code. And add 4 new PV interfaces for the host to
to access the idt/gdt of the guest vcpu. The input parameter is passed
via the per cpu pkvm_pv_param page.

Once the guest state is protected, the idt/gdt is not allowed to be
accessed via these PV interfaces.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>