kvm: pkvm: Set guest state protection flags for pVM
Enable the has_protected_state for the pVM when creates guest. After
returning from the vcpu_run PV interface, set the guest_state_protected
to indicate that the pVM's vcpu state is protected by the pkvm
hypervisor after the vcpu has started running. This is because the host
still pre-configures pVM's vcpu state for booting so the pkvm hypervisor
enforces the protection after that. Once the host doesn't need to do so
then the guest_state_protected flag can be set earlier.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>