pkvm: vmx: Share pVM vcpu state to the host for handling vmcall
Some of the vmcall vmexits will be handled by the host KVM, which
requires to access the corresponding guest registers. Share the vcpu
RAX/RBX/RCX/RDX/RSI according to the vmcall protocol before switching
to the host. Meanwhile accept the vmcall handling result in the RAX
shared from the host after switching to the pkvm hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>