pkvm: x86: Add sync_vcpu_state_pre_switch
Add the new callback sync_vcpu_state_pre_switch in the pkvm_x86_ops for
the pkvm hypervisor to sync architecture specific vcpu state before
switching to the host. For the vmx, the exit related data should be
shared which will be used by the host to identify the exit reason and
handle the vmexit accordingly.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>