pkvm: vmx: Handle CR vmexit
Add code to handle CR vmexit. The CR0/CR4 can be handled by the pkvm
hypervisor by porting from the linux KVM/VMX code. Changing the rule
bits in the CR0/CR4 can result in mmu reset/init. Currently this is done
via setting a pending request in the pkvm_run page and let the host KVM
to do the mmu reset/init. This should be revisited when PV EPT method is
ready. Accessing CR3 won't cause vmexit as the pkvm hypervisor enables
unrestricted guest and disabled the CR3 vmexit. The CR8 exit is also
disabled as the pkvm hypervisor requires TPR shadow.

The PAE emulation is not supported yet by the pkvm hypervisor as it
requires to access guest CR3 page to get the PDPTR which is not
supported.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>