pkvm: x86: Add request HOST_HANDLE_EXIT
Add reqs_to_host to represents the requests the pkvm hypervisor made for
the host to handle. So any request bit set by the pkvm hypervisor can
make it break the vcpu run loop and return back to the host to handle
this.

The new request HOST_HANDLE_EXIT represents the host KVM is required to
handle the vmexits after returning back from the vcpu_run PV interface.
With this, it is not necessary for the vcpu_run PV interface to return
fastpath_t any more.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>