kvm: pkvm: Use PV interface to inject exception for the npVM
Add inject_exception PV interface for the host to inject exceptions to
the npVM. This would make the host KVM being able to emulate more
exceptions for the npVM. Host injecting exception to the pVM is still
not allowed. Do KVM_BUG_ON for the pVM to cache these cases for the
future investigation.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>