kvm: pkvm: Use PV interface to cancel injection
Add cancel_injection PV interface for the host to cancel an injection.
For the npVM, this PV interface can cancel an injected nmi, irq or
exception. But for the pVM, only an nmi or irq can be cancelled as the
pVM's exception is managed by the pkvm hypervisor. The detailed
cancelled event information is saved in the shared kvm_vcpu structure.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>