pkvm: vmx: Handle the NM exception
With passthrough the XFD MSR, NM exception will cause vmexit to let the
hypervisor to save the XFD error timely. As the FPU switching is managed
by the host, the XFD error saving is be done by the host. Meanwhile
injecting the NM exception back to the guest can be done by the pkvm
hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>