REVERTME: pkvm: vmx: Handle the MSR_IA32_XFD MSR vmexit
The MSR_IA32_XFD handling in vmx_set_msr is skipped as currently the FPU
switching is still managed by the host. So the MSR_IA32_XFD emulation is
forwarded to the host to handle.

The host will set the xfd_no_write_intercept to indicate the XFD can
passthrough to the guest. And before entering the guest, the pkvm
hypervisor will update the MSR interception and exeption bitmap
accordingly.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>