kvm: pkvm: Initialize pi_desc
The host response to inject interrupts to the guest. The pi_desc is the
key structure for the host to inject interrupts via the posted interrupt
mechanism. Its physical address is used for the POSTED_INTR_DESC_ADDR in
the VMCS by the pkvm hypervisor. Initialize the pi_desc when reset vcpu.

Fixes: 610ea8af09ed ("kvm: pkvm: Use PV interface to do vcpu reset")
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>