kvm: pkvm: Add update_cpuid_runtime PV interface
Add update_cpuid_runtime PV interface for the host to ask the pkvm
hypervisor to update the cpuid at the runtime. Currently the cpuid
updating is only happened according to the apic base changes emulated by
the host. So even with this PV interface, the host still cannot
arbitrarily change the cpuid entries for the pVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>