pkvm: x86: Enhance the debug registers switching
Before entering the vcpu run loop, save the current debug registers for
the host as they may be overwritten to the guest's value. And restore
the debug registers for the host after exiting the vcpu run loop.

The guest debug registers are loaded before vcpu running if it is
necessary. And syncing the dirty debug registers for the guest inside
the vcpu run loop so that the up-to-date dr variables in the vcpu->arch
can be used.

Fixes: a916fa4f6f1b ("pkvm: x86: Switch the debug registers")
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>