pkvm: x86: Setup fpu kernel and user config for the pkvm hypervisor
To support the pkvm hypervisor doing the FPU isolation for the pVM, add
the fpu_kernel_cfg and fpu_user_cfg symbols to the pkvm hypervisor which
will be used by the pkvm hypervisor to initialize the guest fpstate. The
fpu configs are initialized according to the corresponding host symbols
before the deprivilege the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>