pkvm: x86: Add new PV interface to reallocate fpstate
With expoing the FPU dynamica feature via the cpuid, the fpstate
allocated when creating the vcpu may not be sufficient for the guest.
As the pVM's FPU state is managed by the pkvm hypervisor while the
npVM's FPU state is managed by the host, re-allocating the fpstate is
only necessary for the pVM, and should be done before adding the new
cpuid entries to the pkvm hypervisor. This is only permitted before the
pVM has started running. The old fpstate physical addresses are returned
back to the host via the pkvm memcache.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>