pkvm: x86: Use memcache to teardown the old cpuid entries memory
The vcpu_after_set_cpuid PV interface will add a new cpuid entries to
the pkvm hypervisor if it is allowed. In this case, the previous cpuid
entries memory physical addresses will be returned back to the host. And
such memory is freed with the size of the new cpuid entries, which is
incorrect. To resolve this, the pkvm hypervisor returns such memory
physical addresses via the pkvm memcache format so that the size will
also be returned to the host.

Fixes: 241210f173f0 ("pkvm: x86: Add vcpu_after_set_cpuid PV interface")
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>