x86/paravirt: Extend pv_mmio_ops to support pci mmcfg read and write
Add the pci mmcfg reading and writing callbacks to the pv_mmio_ops so
that can provide platform specific implementations, e.g., an enlightened
kernel can use hypercalls to access the pci mmcfg. The callbacks are
defined to use volatile to qualify the input parameter void __iomme
*pos. Although doing so can result in checkpath warnings, this makes it
possible to patch these callbacks with the read/write wrappers, which is
done by the next patch.

The default implementation is to use "mov" instruction with %eax to
satisfy the AMD CPUs mentioned in the comment of mmio_config_xxx.

Suggested-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Grzegorz Jaszczyk <jaszczyk@google.com>