pkvm: x86: Add percpu fpu
Add a percpu fpu to facilitate the FPU isolation. The percpu fpu is
implemented by leveraging the pcpu_hot->current_task->thread.fpu and
initialized with the initial fpstate, which is used to restore the FPU
to the initial state to isolation the FPU from the host for the pVM. The
permission is set to the maximum features to allow the guest to use all
the supported features.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>