pkvm: x86: Add support to enable guest xfd features
Enable the guest xfd features if the vcpuid expose the feature to the
guest. The fpstate will be reallocated for the pkvm hypervisor before
the cpuid entries are ready. So enabling the xfd feature for the guest
is just to update the xfeatures and xfd fields accordingly if the
fpstate size is sufficient.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>