ANDROID: pkvm: x86: Enforce secure pvmfw bootstrap
When a pVM starts with pvmfw, enforce starting it in 32-bit protected
mode with paging disabled (with the assumption that it is only for
bootstrap, and pvmfw itself will enable paging and switch to 64-bit long
mode as early as possible), similarly to TDX.

The reason for doing that is that 64-bit mode requires paging enabled,
and we cannot trust the initial page tables in the guest memory prepared
by the host VMM (e.g. we need to be sure that the pvmfw entry address in
RIP is not mapped to some other address), so pKVM itself would need to
enforce the initial page tables somewhere in the guest memory, which
would complicate memory management and require a memory layout contract
between pKVM, the VMM and pvmfw.

Of course the initial segment descriptors in the GDT prepared by the
host cannot be trusted either, but we don't need to enforce the initial
GDT in memory, we can just enforce the cached segment descriptors via
VMCS. The CPU will use those cached descriptors instead of GDT until the
guest reloads a segment register (so another assumption is that pvmfw
will avoid reloading segment registers before it sets up its own GDT).
Set the initial GDTR to not point to a valid GDT, so that if the GDT in
memory is still accidentally used too early, it generates a #GP.

While we're at it, since we cannot trust the initial IDT either, set the
initial IDTR to an invalid value as well, so any early exception (#GP in
the above case or any else) will result in a triple fault, rather than
executing an untrusted exception handler prepared by the host.

Bug: 418802144
Bug: 393356668
Test: with crosvm hacked to set up broken initial CPU state (e.g.
CpuMode::FlatProtectedMode with a non-zero CS base), a VM without pvmfw
immediately crashes but a pVM with pvmfw boots.
Upstream-Task: 402758258

Change-Id: Ic81b5792813a9f429692f0a695947ca9136f24c3
Signed-off-by: Dmytro Maluka <dmaluka@google.com>