pkvm: x86: Prevent the vcpu loading race
The host would load the same vcpu on different pCPU at the same time if
the host wants to do that. And this was not properly handled by the pkvm
hypervisor due to the vcpu->cpu is not updated atomically.

To prevent this, use cmpxchg to check and update the vcpu->cpu
atomically so that when one pCPU updates vcpu->cpu then another pCPU
can notice this and stop the illegal loading.

Fixes: cff23c5293fd ("pkvm: x86: Add vcpu_load PV interface")
Fixes: 685a7366dd74 ("pkvm: x86: Add vcpu_put PV interface")
Reported-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>