ANDROID: pkvm: x86: Clean up pre-configuring vCPU by host
The comments in pkvm_vcpu_update_state_from_host() and
is_kvm_vcpu_accessible() are a bit misleading, since the host needs to
pre-configure some of the initial state of the boot vCPU anyway.
For now we allow the host to pre-configure almost anything, as long as
the vCPU has not started running. If the pVM is running with pvmfw
and thus some of the initial state needs to be enforced by pKVM in order
to start pvmfw securely, pKVM enforces that just before running the vCPU
first time, overriding what the host has set before. (Before that, pKVM
does not even know if the pVM will run with pvmfw or not, since the host
VMM may set the pvmfw address via ioctl at any time before running the
first vCPU.)

So update the comments accordingly.

Besides the comments, slightly modify the actual behavior, and define it
more precisely:

- Allow the host to pre-configure any GPRs except R15. This gives us
  flexibility of using various boot protocols between the VMM and pVM
  and/or pvmfw. The R15 is reserved by pKVM itself for potential future
  extensions (in case pKVM itself will want to pass some data to pVM)
  and is set to 0 (define this value as "pKVM boot protocol version"),
  similarly to pKVM-ARM which reserves arm64's X15 for that.

- For secondary vCPUs, don't allow the host to set GPRs (since the host
  doesn't need that), only allow it to set the RIP (temporarily, as we
  are still relying on INIT/SIPI emulation by the host).

Bug: 418802144
Test: boot pVM with pvmfw with multiple vCPUs
Upstream-Task: 402758258

Change-Id: If6602484e2d9a7a15cbdaf6126b397e18d74162e
Signed-off-by: Dmytro Maluka <dmaluka@google.com>