ANDROID: pkvm: x86: Disable FRED for the pkvm hypervisor
Disable FRED for the pkvm hypervisor if it is enabled by the host.
There is no too much benifit for the pkvm hypervisor to use the FRED
event delivery as the NMI is the only event expected to be received
by the pkvm hypervisor. The exceptions are not expected to be happened
in the pkvm hypervisor and all hardware interrupts will directly go to
the host. Meanwhile, enabling the FRED in the pkvm hypervisor will result
in additional FRED MSRs switching overhead. So keep the FRED being
disabled in the pkvm hypervisor.

Bug: 425822773
Upstream-Task: 402758258

Change-Id: Ie6e632067a97580bb99fe746b6985f4f07509a55
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>