ANDROID: pkvm: x86: Refactor the exception/NMI handlers
Enhance the IDT asm code to push the CPU registers in the order of the
structure pt_regs format, so that more CPU registers information can be
passed to the handlers. Meanwhile also pass the vector number and
has_error_code flag to the handler so that the handle_nmi can be called
according to the vector number as a static function. To generate
dedicated IDT entry for each vector, introduced a new header
GEN-for-each-exc.h to facilitate this. Each vector from 0 ~ 31 including
the NMI has its own IDT entry, which is used to initialize the pkvm's
IDT.

As the pkvm hypervisor doesn't expect to receive exceptions if
everything works fine and only cares about the NMI, it is not necessary
to use the host IDT in the debug mode. Meanwhile, the host IDT is
invalid when the FRED feature is enabled and keeping use the host IDT in
debug mode can trigger kernel panic, so use the pkvm's own IDT in the
debug mode as well. In case any exception happened, the exception
handler in the pkvm hypervisor can also print the logs to show detailed
RIP information, exception number and the error code which are helpful
for debugging.

[ Prabu: root caused kernel panic with fred=on and provided a fix ]

Bug: 425822773
Upstream-Task: 402758258
Ignore-Checkpatch: %px is only triggered in non-secure debug mode

Change-Id: I61d7a5616f947fd4dd0221d21470f442e4d6259c
Signed-off-by: Renganathan, Prabu <prabu.renganathan@intel.corp-partner.google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>