ANDROID: pkvm: x86: Don't use e820__get_entry_type() for pvmfw check
e820__get_entry_type() is buggy by design (see [1]): it returns the e820
type of the given memory range, but if this range is not covered by a
single e820 region but is covered, for example, by several e820 regions
of different types, it doesn't check if they all have the same type, it
just returns the type of the last region.

As a result, it gives a false positive if pvmfw is only partially
covered by a E820_TYPE_RESERVED region.

So use e820__mapped_all() instead.

Bug: 350694931
Test: pkvm init fails and pvmfw is cleared if only 2nd half of the pvmfw
region is in e820 reserved memory.
Upstream-Task: 402758258

[1] https://lore.kernel.org/all/aAKCaQRYmjNaz175@google.com/

Change-Id: I19735f90552cc338df5527838b4e3831b9086a92
Signed-off-by: Dmytro Maluka <dmaluka@google.com>