pKVM: x86: Rename all the pKVM hypervisor symbols
The pKVM hypervisor will be isolated from the linux kernel thus it
should not dynamically link to any linux kernel's symbols, but compile
the relevant code as part of the pKVM hypervisor. Some of the functions
in memset_64.c, x86.c and vmx.c are used by the pKVM hypervisor, e.g.,
memset(), kvm_inject_gp(), vmx_inject_exception(), vmxxx_error() and so
on. As these functions will also be required to support guest VMs later
so don't bother creating pKVM's owner functions for the host vm but
directly including these file, with stripping the unnecessary symbols
off from these files via using the define __PKVM_HYP__.

However, this will bring the symbol name conflicts because the linux
kernel also have the same file compiled. To resolve this, rename all the
pKVM hypervisor symbols via adding '__pkvm' suffix to avoid the
conflicts, except for the compiler generated symbols __x86_return_thunk
and __x86_indirect_XXX to satisfy the objtool and boot time patching
(the subsequent patch will introduce pKVM's own retpoline to support the
pKVM text boot time patching). The reason of adding suffix rather than
prefix is to not conflict with other prefixes like __cfi.

Another benefit of adding a suffix is that this can help caching the
incorrect linux kernel symbols usage in the pKVM hypervisor at the
linking time rather than at the running time, which is easier to debug
and fix.

As the pKVM hypervisor code doesn't export any symbol for the kernel to
use, by pass the  __EXPORT_SYMBOL for the pKVM hypervisor code to
strip off the export related symbols.

To strip off the unused kernel symbols from the pKVM, introduced
undef.h. This is included by each pKVM source code to undefine certain
kernel configs, so that the symbols enabled by these kernel configs can
be stripped off, e.g, #undef CONFIG_PRINTK to remove printk symbols. The
kernel CONFIG_BUG and CONFIG_GENERIC_BUG also introduced some unused
symbols, however undefining these kernel configs can result in the
following warnings from the resolve_btfids tool:

    WARN: multiple IDs found for 'task_struct': 395, 80655 - using 395
    WARN: multiple IDs found for 'inode': 195, 80666 - using 195
    WARN: multiple IDs found for 'module': 83, 80710 - using 83
    WARN: multiple IDs found for 'file': 667, 80759 - using 667
    WARN: multiple IDs found for 'vm_area_struct': 371, 80803 - using 371
    WARN: multiple IDs found for 'seq_file': 769, 80860 - using 769
    WARN: multiple IDs found for 'path': 842, 80904 - using 842
    WARN: multiple IDs found for 'bpf_prog': 911, 81006 - using 911
    WARN: multiple IDs found for 'bpf_map': 979, 81019 - using 979
    WARN: multiple IDs found for 'cgroup': 642, 81093 - using 642
    WARN: multiple IDs found for 'bpf_link': 1286, 81204 - using 1286
    WARN: multiple IDs found for 'task_struct': 395, 111581 - using 395
    WARN: multiple IDs found for 'inode': 195, 111592 - using 195
    WARN: multiple IDs found for 'module': 83, 111636 - using 83
    WARN: multiple IDs found for 'file': 667, 111685 - using 667
    WARN: multiple IDs found for 'vm_area_struct': 371, 111729 - using 371
    WARN: multiple IDs found for 'seq_file': 769, 111806 - using 769
    WARN: multiple IDs found for 'path': 842, 111850 - using 842
    WARN: multiple IDs found for 'bpf_prog': 911, 111952 - using 911
    WARN: multiple IDs found for 'bpf_map': 979, 111965 - using 979
    WARN: multiple IDs found for 'cgroup': 642, 112039 - using 642
    WARN: multiple IDs found for 'bpf_link': 1286, 112150 - using 1286

The reason is that undefining these kernel configs makes the layout of
the above kernel structures in the hypervisor different from their
layout in the host. Even if that is not harmful for pKVM itself, it at
least may cause CONFIG_DEBUG_INFO_BTF itself to work incorrectly, as it
makes the resolve_btfids tool confused as to which of the multiple
layouts of the same struct to use, when adding the BTF type info to the
kernel image. So directly modify the corresponding those kernel's header
files to use them in the hypervisor code.

[ Dmytro: Remove symbols enabled by CONFIG_BUG and CONFIG_GENERIC_BUG ]

Co-developed-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>