pkvm: x86: Implement a pgtable mm_ops for early alloc
The pkvm hypervisor will use the early alloc to allocate memory pages for
its own purpose, e.g., create its own MMU page table. Implement a pgtable
memory management ops for this purpose. As the early allocated memory
pages cannot be freed, it is not necessary to support the get_page and
the put_page callbacks. The page count is also meaningless.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>