pkvm: x86: Add pkvm_constants.c to expose the sizes of pkvm only data types
The structure pkvm_page is a pkvm only data type, which is used by the
pkvm buddy allocator. However, the host VM still needs to know the size
of the structure pkvm_page so that it can calculate how much memory
needs to be reserved for the pkvm_page structures. To avoid exposing
the such pkvm only data type to the host VM, introduce pkvm_constants.c
which can be compiled as a pkvm_constants.h header file, which contains
the constant sizes of certain pkvm only data types.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>