pKVM: x86: Reserve memory pages for the host mmu
Reserve memory pages for the host mmu at the early boot. The reserved
size is calculated similarly with the pKVM mmu. The difference is that
the host mmu will also include the MMIO mapping for the host VM, thus
counting more memory pages for the 2GB MMIO hole.

MMIO regions also exists in the high-end physical address space which is
not able to know the size precisely during the early boot. Therefore,
instead of specifically reserving memory pages for these MMIO regions,
share the reservation for the 2G ~ 4G MMIO region. While this approach
cannot guarantee that memory page allocation for the required MMIO
mappings will always success, it is feasible because the reservation for
2G ~ 4G MMIO region is based on the worst case (4K page size) while in
practice the MMIO region will be mapped using the possible largest page
size(e.g. 1G/2M), which significantly reduces the memory consumption and
makes sharing possible.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>