pKVM: VMX: Handle the deprivileged host vmexit
Although the VMCS is configured to minimize the vmexits for the
deprivileged host, there are still some vmexits cannot be avoid.

1. The CPUID/XSETBV vmexit is handled by re-executing the same
instruction on the native and returning the corresponding result to the
host.

2. The CR vmexit only handles mov-to-cr4 to prevent the host really
controlling the VMX on hardware.

3. The MSR vmexits are handled by injecting the #GP to the host VM as
currently only invalid MSR accessing will cause vmexits. Reuse the KVM
source code to queue and inject exception.

The other types of vmexits should not happen at this moment.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>