DEBUG: pkvm: x86: Add log printing capability for debug
Introduce the kernel config option CONFIG_PKVM_X86_DEBUG to provide
support for the pKVM hypervisor to print logs by using the linux kernel
symbols. This is achieved by the pKVM hypervisor cloning the host VM's
mmu mapping to its own mmu, as well as reusing some of the host VM's CPU
state as the pKVM hypervisor's CPU state, including the percpu area.

With this, the pKVM hypervisor can use printk/trace_printk symbols to
print logs. As the symbols used by the pKVM hypervisor will be renamed
during compiling, alias these debug symbols.

As enabling CONFIG_PKVM_X86_DEBUG will allow the pKVM hypervisor to use
the linux kernel symbols, which are untrusted code as can be modified
easily by the host, the security of the pKVM hypervisor cannot be
guaranteed. So suggest to enable this only for debugging purpose.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>