pkvm: vmx: Intercept writing to some X2APIC MSRs
The lapic is used by the pkvm in x2apic mode, to kick remote CPU via
sending INIT signal with the remote apic id. So it is important for the
lapic being always in the x2apic mode with its apic id unchanged. To
satisfy this, intercept writing to the MSR_IA32_APICBASE and APIC_ID MSR
to guarantee this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>