pKVM: x86: Add vendor host mmu init ops
Add the host mmu instance to represent the 2nd-level page table for the
host VM. The host_mmu is a x86 generic instance. Different vendors may
have different initializations, e.g., initialize the host_mmu as an EPT
on Intel platforms.

Add the host_mmu_init callback to the pkvm_init_ops, for vendors to
initialize the host mmu with the memory pool divided from the reserved
memory area during the finalize.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>