pkvm: x86: Switch to use buddy allocator for the hyp mmu
The hyp mmu was initialized by using the early allocator. After the
pkvm_vmemmap is backed in the hyp mmu, the buddy allocator is ready to
use. So load the pkvm hypervisor's own mmu, and re-initialize the hyp
mmu page table with using buddy allocator.

Since some of hyp mmu pages were allocated by the early allocator which
is not able to initialize the page refcount stored in the pkvm_page
structure, fix this by walking the hyp mmu page table and initialize
the page refcount for each present page table pages.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>