pkvm: x86: Add page table lookup support
Add pkvm_pgtable_lookup() to support look up the mapping information of
a virtual address in a page table. If the leaf is present, return the
physical address bits and property bits in the pte, as well as the level
of the present leaf. If the leaf is non-present, returns INVALID_PAGE as
the physical address bits, 0 as the property bits, as well as the level
of the non-present leaf.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>