pKVM: x86: Add helper to find the address range for a given phys addr
Add helper pkvm_find_addr_range which can find out the address range
that contains the given physical address. The address range would be
either a memory range if the given physical address is memory, or a hole
(which is between memory ranges and usually can be used as the MMIO)
range if the given physical address is not memory, which is indicated by
the return value.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>