pKVM: x86: Setup percpu area for the pKVM hypervisor
The percpu mechanism is useful for the pKVM hypervisor to manage percpu
data. However the pKVM hypervisor cannot use the linux kernel's percpu
memory for security reason. And having pKVM's own percpu memory is also
helpful for the pKVM to reuse some of the linux kernel's code.

Define a pKVM percpu section to gather all the pKVM percpu symbols. More
memory pages are reserved for each CPU, and will be used as the
__per_cpu_offset and this_cpu_off. The __per_cpu_offset will be used as
pKVM's GS_BASE when enables the isolation by the later patches.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>