KVM: pVMX: Add initial pKVM host interface
Add the initial implementation of the pKVM host, which serves as a
middle layer between the host KVM and the pKVM hypervisor. An empty
pkvm_host_x86_ops structure is introduced, which will be extended to
provide the necessary operations for the host KVM to interact with
pKVM to run guest VMs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>