pKVM: x86: Expose PKVM_VMX_VM_SIZE for vmx host
Expose PKVM_VMX_VM_SIZE to allow the VMX host to know the size of
a VM instance which is managed by the pKVM hypervisor, so that no need
to expose struct pkvm_vm which is internal to the pKVM hypervisor.

Defining and exposing for the size for the VMX host is sufficient as
currently VMX host is the only supported host by the pKVM hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>