KVM: pVMX: Implement vcpu_create operation
Enable the host KVM to create a vCPU instance which is managed by the
pKVM hypervisor via the vcpu_create PV interface. Before that, the VMX
host needs to allocate enough memory pages for the VMX hardware to use,
i.e., VMCS page, MSR bitmap, PML page and VE page, which will be donated
to the pKVM hypervisor. Additionally, memory pages must be allocated for
vCPU instance, as well as the FPU state memory.

The vcpu_create PV interface returns a handle that allows the host to
identify the corresponding vCPU instance managed by the pKVM hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>