pKVM: x86: Expose PKVM_VMX_VCPU_SIZE
Expose PKVM_VMX_VCPU_SIZE to allow the VMX host to know the size of a
vCPU instance which is managed by the pKVM hypervisor, so that no need
to expose struct pkvm_vcpu which is internal to the pKVM hypervisor.

Defining and exposing the size for the VMX host is sufficient as
currently VMX host is the only supported host by the pKVM hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>