KVM: pKVM: Implement disable_virtualization_cpu operation
Enable the host KVM to disable virtualization features which were
required for running guest VMs on a specific CPU via the
disable_virtualization_cpu PV interface.

The pKVM hypervisor doesn't support disabling VMX via this PV interface.
As a result, the CPU will remain in VMX non-root mode during shutting
down or rebooting if there is no hardware level reset. In this case, if
INIT-SIPI are send, such events will cause vmexits to the pKVM
hypervisor rather than the operations associated with these events.
Handling such vmexits is not supported by the pKVM hypervisor. Put a
TODO for this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>