KVM: x86/mmu: pKVM: Skip kvm_arch_flush_shadow_all()
With regular KVM, kvm_arch_flush_shadow_all() is responsible for
unmapping all pages mapped in the guest page tables and (perhaps more
importantly) freeing memory used for the guest page tables when
destroying the VM.

With pKVM, the pKVM hypervisor takes care of that when destroying the
VM anyway [*], so no need for extra actions on the host side for that
(especially given that for protected VMs, pKVM will not allow to unmap
pages before destroying the VM anyway).

That said, this means that pKVM takes quite a lot of time to handle
the vm_destroy hypercall. We may optimize that in the future, similarly
to pKVM-ARM in commit 49f39131dff2 ("ANDROID: KVM: arm64: Refactor the
guest teardown path").

[*] FWIW, unlike KVM, pKVM does not unmap guest pages when destroying
the VM (there is no need to, since guest vCPUs are no longer running),
it just frees the guest page table. That's an implementation detail
though.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>