pKVM: x86: Add vcpu_load PV interface
Add the vcpu_load PV interface to load a guest vCPU. The reference count
of the corresponding pkvm_vcpu instance is incremented to prevent it
from being freed by the host. The guest vCPU doesn't need to be loaded
on the physical CPU immediate, which can be delayed until it is needed.
A percpu cur_guest_vcpu is introduced to represent the current loaded
guest vCPU.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>