pKVM: x86: Redefine KVM print macros for the pKVM hypervisor
The KVM print macros may be included in the code which are leveraged
by the pKVM hypervisor. Overridden these print macros to use the "pkvm"
prefixes, and avoid printing current task information which is not
supported in the pKVM hypervisor.

The symbols enabled by CONFIG_DYNAMIC_DEBUG/_CORE cannot be removed by
undefine these kernel configs in the undef.h due to the same reason with
the CONFIG_BUG/CONFIG_GENERIC_BUG. Directly modify the corresponding
kernel's header files to remove these symbols for the pKVM hypervisor.

[ Dmytro: Remove symbols enabled by CONFIG_DYNAMIC_DEBUG/_CORE ]

Co-developed-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>