pKVM: x86: Add load_mmu_pgd PV interface
Add load_mmu_pgd PV interface for the host to load the guest mmu via the
vendor-specific operation. Currently the guest mmu for both pVM and npVM
are still managed by the host, thus this PV interface loads the guest
mmu from the host. Eventually the guest mmu will be managed inside the
pKVM to achieve pVM's isolation, thus this PV interface will be updated
to load the pKVM managed guest mmu and always accessible to the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>