pKVM: x86: Add vcpu_reset PV interface
Add vcpu_reset PV interface to allow the host to reset guest vCPU for
INIT event via the kvm_vcpu_reset(). It is not necessary to reset guest
vCPU for non-init event by the host as the guest vCPU is reset by the
pKVM hypervisor after it is created.

This PV interface is enforced to ensure a pVM cannot be reset by the
host once the pVM has started running. As the INIT for a pVM will be
handled inside the pKVM hypervisor for the security reason, eventually
this PV interface will be only allowed for npVMs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>