pKVM: VMX: Finalize host ept
Implement the VMX host_mmu_finalize callback. Program the host EPT root
and enable EPT in the VMCS so the next host entry enforces the mappings
prepared by pKVM.

Enable VPID when supported. Derive a nonzero VPID from the host vCPU ID,
because VPID zero is reserved for translations that are not tagged, and
reject IDs that exceed the architectural VPID width.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>