pKVM: x86: Add vcpu_create PV interface
Introduce the vcpu_create PV interface to allow the host to create a
vCPU for a specific guest VM.

As the pKVM may need to access the host's kvm_vcpu structure, share its
memory pags with the pKVM to prevent those pages from being used as
other purposes (e.g., donated to the pKVM or used as guest memory). And
the pkvm_vcpu structure is private to pKVM, so its memory pages are
donated from the host.

Each pkvm_vcpu is assigned a handle within the VM, which is returned
to the host via the vcpu_create interface. A pkvm_vm can support up
to KVM_MAX_VCPUS vCPUs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>