pKVM: x86: Initialize the percpu fpu
Initialize the percpu current_task fpu with the initial fpstate, which is
used to restore the FPU to the initial state when switches from pVMs to
the host VM, so that to isolation pVM's FPU from host. The permission is
set to the maximum features to allow the guest to use all the supported
features.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>