KVM: pVMX: Add new fpstate memory for xfd
Add new fpstate memory via the vcpu_add_fpstate PV interface if the xfd
feature is enabled for pVM. As the pVM's FPU state is managed by the
pKVM hypervisor while the npVM's FPU state is managed by the host,
re-allocating is only necessary for the pVM, and should be done before
adding the new cpuid entries to the pKVM hypervisor.

If the new fpstate memory is added successfully, free the old one.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>