pKVM: x86: Add vcpu_run PV interface
Add vcpu_run PV interface for the host to run a guest vCPU. This PV
interface is always accessible to the host.

To run a guest vCPU, the necessary guest vCPU state will be updated
according to the shared vCPU state owned by the host. If the guest is a
npVM, the guest vCPU state will be updated unconditionally. If the guest
is a pVM, only BSP vCPU state will be updated if the BSP vCPU is not
starting to run to guarantee the guest state security. The RIP of AP
vCPU is still allowed to be updated if AP vCPU is not starting to run,
as a temporary solution to bring up AP. This will be fixed by a secure
INIT-SIPI mechanism which will be introduced later.

Before returning back to the host, the npVM's vCPU state will be shared
with the host unconditionally. The pVM's vCPU state will be shared
according to the different vmexit reasons, which will be implemented by
the following patches.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>