pKVM: x86: Import kvm_skip_emulated_instruction()
Import kvm_skip_emulated_instruction from the KVM for the pKVM to skip
instructions.

If the guest_debug is KVM_GUESTDBG_SINGLESTEP, set the request
HOST_HANDLE_GUESTDBG_SINGLESTEP to let the pKVM hypervisor switch back
to the host to handle this rather than entering the guest. This will not
happen to the pVM as debugging the pVM is not supported.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>